* help_plugin: Fix looking up symbolic linked man pages in gz or bz2 files (bug #19155)
The problem happened because the man file stored the path without specifying
the extension of the target man file on disk. This commit adds check if the
linked file exists and if it doesn't then it checks if there is a gz or bz2
file with the same name. It such file exists it loads it.

git-svn-id: svn://svn.code.sf.net/p/codeblocks/code/trunk@9516 2a5c6006-c6dd-42ca-98ab-0921f2732cef