* CCTest: don't save temporary files, just parse the files specified by the file name.
If you have a test file, e.g. testing/a.cpp, the former way was load the file to wxScintilla control, when start parsing, we first save the contents of wxScintilla control to a temporary file, then parse it, and finally delete the temporary file. With this commit, we don't need to use temporary files, the file can be parsed directly if a file name is specified. The ParserBase class can directly load the file from hard disk for parsing.

git-svn-id: svn://svn.code.sf.net/p/codeblocks/code/trunk@9970 2a5c6006-c6dd-42ca-98ab-0921f2732cef